requirements.txt
python-dotenv==0.21.1
SQLAlchemy==2.0.4
pyTelegramBotAPI==4.10.0
requests==2.26.0

flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3